requirements.txt
pandas==2.2.2
datasets==2.20.0
transformers==4.41.2
sentence-transformers==2.7.0
torch==2.3.0
ipywidgets==8.1.2
matplotlib==3.8.4
seaborn==0.13.2
hydra-core==1.3.2
numpy==1.26.4